add calculating scores

+++ models/game_model.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import Column, ForeignKey, Boolean, Integer
+from sqlalchemy import Column, ForeignKey, Boolean, Float, Integer
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 
@@ -13,7 +13,7 @@ class Game(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     started = Column(Boolean)
     finished = Column(Boolean)
-    score = Column(Integer)
+    score = Column(Float)
     offset = Column(Integer)
     user_id = Column(UUID(as_uuid=True), ForeignKey("users.id"))
     quiz_id = Column(UUID(as_uuid=True), ForeignKey("quizzes.id"))
